requirements.txt
Flask
google-cloud
